requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1      
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2        
async-timeout==4.0.3  
attrs==23.1.0
backoff==2.2.1        
bcrypt==4.1.2
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.3
decorator==4.4.2
Deprecated==1.2.14
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
emoji==2.9.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.103.2
ffmpeg==1.4
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
importlib-resources==6.1.1
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==28.1.0
langchain==0.0.339
langdetect==1.0.9
langsmith==0.0.75
llvmlite==0.41.1
lxml==4.9.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mmh3==4.0.1
monotonic==1.6
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.3.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0       
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
Pillow==10.1.0
posthog==3.1.0
proglog==0.1.10
protobuf==4.25.1
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pypdf==3.17.1
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.4.1
python-box==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruff==0.1.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
unstructured==0.11.2
urllib3==1.26.18
uvicorn==0.23.2
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0